add models for rating and comment
